Alex remote commit

+++ client_code/components/GridView.py
@@ -330,12 +330,12 @@ class GridView:
 
 
     def row_selected(self, args):
-        print('row_selected', args)
+        # print('row_selected', args)
         self.grid.element.querySelector(f'.e-toolbar .e-toolbar-item[title="Delete"]').style.display = 'inline-flex'
     
     
     def row_deselected(self, args):
-        print('row_deselected', args)
+        # print('row_deselected', args)
         if not self.grid.getSelectedRecords():
             self.grid.element.querySelector(f'.e-toolbar .e-toolbar-item[title="Delete"]').style.display = 'none'
     
@@ -393,25 +393,28 @@ class GridView:
             })
     
     
-    def delete_selected(self, args):
+    def delete_selected(self, args, persist=True):
         print('delete_selected')
+        if self.confirm_dialog:
+            self.confirm_dialog.hide()
+            self.confirm_dialog.destroy()
+            self.confirm_dialog = None
+            
         for grid_row in self.grid.getSelectedRecords() or []:
             print('Delete row', grid_row)
-            db_row = self.grid_class.get(grid_row.uid) if grid_row.uid else None
-            if db_row is not None:
-                db_row.delete()
             if grid_row.uid:
                 print('uid', grid_row.uid)
                 self.grid.deleteRecord('uid', grid_row)
             else:
                 print('no uid')
                 self.grid.dataSource.remove(grid_row)
-                self.grid.refresh()
-        if self.confirm_dialog:
-            self.confirm_dialog.hide()
-            self.confirm_dialog.destroy()
-            self.confirm_dialog = None
-        # self.grid.refresh()
+        self.grid.refresh()
+
+        if persist:
+            for grid_row in [x for x in self.grid.getSelectedRecords() or [] if x.uid]:
+                db_row = self.grid_class.get(grid_row.uid) if grid_row.uid else None
+                if db_row is not None:
+                    db_row.delete()
 
 
     def update_grid(self, data_row, add_new):
